pyprojects/secretflow-doctools/pyproject.toml
[project]
authors = [{name = "Tony Wu", email = "tonywu6@protonmail.com"}]
dependencies = [
  "aiofiles~=23.1.0",
  "asyncstdlib>=3.10.8",
  "click>=8.1.5",
  "importlib-resources>=6.0.0",
  "loguru>=0.7.0",
  "more-itertools>=9.1.0",
  "pexpect>=4.8.0", # "pydantic-settings>=2.0.2",
  "pydantic>=1.10.12,<2.0.0",
  "python-dotenv~=1.0.0",
  "ruamel-yaml>=0.17.32",
  "sphinx>=6.2.1",
  "tabulate>=0.9.0",
  "termcolor>=2.3.0",
  "tqdm>=4.65.0",
  "twisted>=22.10.0",
  "typing-extensions>=4.7.1",
  "urllib3<2.0.0", # urllib3>=2.0.0 breaks requests-unixsocket
  "requests>=2.31.0",
  "requests-unixsocket>=0.3.0",
  "watchfiles>=0.20.0",
  "pathspec>=0.11.1",
]
description = "Documentation toolchain for SecretFlow"
dynamic = ["version"]
name = "secretflow-doctools"
readme = "README.md"
requires-python = ">=3.8.15, <3.11"

[project.scripts]
secretflow-doctools = 'secretflow_doctools.__main__:cli'

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "package.json"
# regex to match "version" in package.json
pattern = '"version": "(?P<version>[^"]+)"'

[tool.hatch.build.targets.wheel]
sources = ["src"]

[tool.hatch.build]
artifacts = ["**/node_modules/mdserver/dist"]
directory = "dist"
only-include = ["src", "package.json"]

[tool.hatch.build.targets.custom]
path = 'scripts/hatch_build.py'
sources = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
